Set basic user permissions @ posts.view

+++ DAMSBlog/posts/views.py
@@ -1,6 +1,6 @@
 from django.contrib import messages
 from django.shortcuts import render, get_object_or_404, redirect
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.core.paginator import Paginator, PageNotAnInteger, EmptyPage
 
 from urllib.parse import quote_plus
@@ -12,6 +12,9 @@ from .forms import PostForm
 
 
 def posts_create(request):
+    if not request.user.is_staff or not request.user.superuser:
+        raise Http404
+
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
@@ -60,6 +63,9 @@ def posts_list(request):
 
 
 def posts_update(request, id=None):
+    if not request.user.is_staff or not request.user.superuser:
+        raise Http404
+
     instance = Post.objects.get(id=id)
     form = PostForm(request.POST or None, request.FILES or None, instance=instance)
     if form.is_valid():
@@ -77,6 +83,8 @@ def posts_update(request, id=None):
 
 
 def posts_delete(request, id=None):
+    if not request.user.is_staff or not request.user.superuser:
+        raise Http404
     instance = get_object_or_404(Post, id=id)
     instance.delete()
     messages.success(request, "Successfully deleted")
